requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-auth==2.24.0
google-auth-oauthlib==1.1.0
gspread==5.12.2
idna==3.6
MouseInfo==0.1.3
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.0
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyTelegramBotAPI==4.14.0
pytesseract==0.3.10
pytweening==1.0.7
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
soupsieve==2.5
telebot==0.0.5
urllib3==2.1.0

